first round of PR comments
