+++ io/CMakeLists.txt
@@ -10,6 +10,8 @@ message(STATUS "Building 'detray::io' component")
 # Create the "main I/O library".
 detray_add_library( detray_io io )
 
+add_subdirectory ( payload )
+
 # Include all active components.
 if( DETRAY_IO_CSV )
    add_subdirectory( csv )
