+++ io/json/include/detray/io/json_geometry_io.hpp
@@ -9,6 +9,7 @@
 
 #include <array>
 
+#include "detray/io/io_payload.hpp"
 #include "detray/io/json_algebra_io.hpp"
 #include "detray/io/json_defs.hpp"
 #include "detray/io/json_material_io.hpp"
@@ -19,23 +20,6 @@
 /// JSON I/O extension
 namespace detray {
 
-/// @brief  A payload object for masks
-struct mask_payload {
-    enum class mask_type : unsigned int {
-        annulus2 = 0u,
-        cuboid3 = 1u,
-        cylinder2 = 2u,
-        cylinder3 = 3u,
-        line = 4u,
-        rectangle2 = 5u,
-        ring2 = 6u,
-        single3 = 7u,
-        trapezoid2 = 8u
-    };
-    mask_type type = mask_type::ring2;
-    std::vector<real_io> boundaries;
-};
-
 void to_json(nlohmann::json& j, const mask_payload& m) {
     j["type"] = static_cast<unsigned int>(m.type);
     j["boundaries"] = m.boundaries;
@@ -45,14 +29,6 @@ void from_json(const nlohmann::json& j, mask_payload& m) {
     m.type = static_cast<mask_payload::mask_type>(j["type"]);
     m.boundaries = j["boundaries"].get<std::vector<real_io>>();
 }
-
-/// @brief  A payload for surfaces
-struct surface_payload {
-    transform_payload transform;
-    mask_payload mask;
-    material_slab_payload material;
-    std::size_t gid;
-};
 
 void to_json(nlohmann::json& j, const surface_payload& s) {
     j["transform"] = s.transform;
@@ -68,12 +44,6 @@ void from_json(const nlohmann::json& j, surface_payload& s) {
     s.material = j["material"];
 }
 
-/// @brief a payload for portals
-struct portal_payload {
-    surface_payload surface;
-    links_payload volume_links;
-};
-
 void to_json(nlohmann::json& j, const portal_payload& p) {
     j["surface"] = p.surface;
     j["volume_links"] = p.volume_links;
@@ -84,18 +54,6 @@ void from_json(const nlohmann::json& j, portal_payload& p) {
     p.volume_links = j["volume_links"];
 }
 
-/// @brief a payload for volume bounds
-struct volume_bounds_payload {
-    enum class volume_bounds_type {
-        cuboid = 0u,
-        cylindrical = 1u,
-        generic_cuboid = 2u,
-        trapezoid = 3u
-    };
-    volume_bounds_type type = volume_bounds_type::cylindrical;
-    std::vector<real_io> values = {};
-};
-
 void to_json(nlohmann::json& j, const volume_bounds_payload& vb) {
     j["values"] = vb.values;
     j["type"] = static_cast<unsigned int>(vb.type);
@@ -105,16 +63,6 @@ void from_json(const nlohmann::json& j, volume_bounds_payload& vb) {
     vb.values = j["values"].get<std::vector<real_io>>();
     vb.type = static_cast<volume_bounds_payload::volume_bounds_type>(j["type"]);
 }
-
-/// @brief a payload for volumes
-struct volume_payload {
-    std::string name = "";
-    transform_payload transform;
-    volume_bounds_payload volume_bounds;
-    std::vector<portal_payload> portals;
-    std::vector<surface_payload> surfaces;
-    links_payload surface_links;
-};
 
 void to_json(nlohmann::json& j, const volume_payload& v) {
     j["name"] = v.name;
@@ -151,13 +99,6 @@ void from_json(const nlohmann::json& j, volume_payload& v) {
         v.surface_links = j["surface_links"];
     }
 }
-
-/// @brief a payload for volumes
-struct detector_payload {
-    std::string name = "";
-    std::vector<volume_payload> volumes = {};
-    grid_objects_payload volume_grid;
-};
 
 void to_json(nlohmann::json& j, const detector_payload& d) {
     j["name"] = d.name;
